+++ src/Exceptions/BringApiException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Tor2r\BringApi\Exceptions;
+
+use Exception;
+
+class BringApiException extends Exception {}